+++ rsync.delta/DeltaWriter.cs
@@ -5,6 +5,7 @@ using System.Diagnostics;
 using System.IO.Pipelines;
 using System.Threading;
 using System.Threading.Tasks;
+using Rsync.Delta.Pipes;
 
 namespace Rsync.Delta
 {
@@ -158,7 +159,7 @@ namespace Rsync.Delta
 
         private async ValueTask<BufferedBlock> BufferBlock(CancellationToken ct)
         {
-            ulong len = _blocks.BlockLength + _pendingLiteralLength;
+            ulong len = (uint)_blocks.BlockLength + _pendingLiteralLength;
             var readResult = await _reader.Buffer((long)len, ct); // fix this cast
             var pendingLiteral = readResult.Buffer.Slice(0, (int)_pendingLiteralLength); // fix
             var currentBlock = readResult.Buffer.Slice((int)_pendingLiteralLength); // fix
